Extract existing enums into files
