+++ src/utils/Firestore.ts
@@ -1,7 +1,8 @@
 import 'firebase/functions';
-import { CallableFunctions, ListName } from './consts';
 import { CurrentFamily, Family, Product, WastedProduct } from '@/types';
 import { productConverter, wastedProductConverter } from '@/types/converters';
+import { CallableFunctions } from './enums';
+import { ListName } from './consts';
 import firebase from 'firebase';
 
 import FieldValue = firebase.firestore.FieldValue;
