+++ src/utils/enums/CallableFunctions.ts
@@ -1,0 +1,4 @@
+export enum CallableFunctions {
+  GetUsersByEmail = 'getUsersByEmail',
+  PredictImage = 'predict'
+}
