+++ src/utils/enums/ErrorCode.ts
@@ -1,0 +1,8 @@
+export enum ErrorCode {
+  InvalidEmail = 'auth/invalid-email',
+  InvalidPassword = 'auth/invalid-password',
+  WrongPassword = 'auth/wrong-password',
+  WeakPassword = 'auth/weak-password',
+  InvalidDisplayName = 'auth/invalid-display-name',
+  UserNotFound = 'auth/user-not-found'
+}
